+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add option to set system properties from command-line
+- Add Maven Enforcer custom rule [#3](https://github.com/nbbrd/heylogs/issues/3)
 
 ### Fixed